requirements.txt
polygon-api-client==1.13.6
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
schedule==1.2.2
pytz==2024.1